requirement.txt
flask
torch
torchvision
transformers
sentencepiece
pillow
